r/aws_glue_ml_transform: Reduce visibility.
